+++ wdl-lint/src/rules.rs
@@ -1,7 +1,7 @@
 //! Module for the lint rules.
 
 mod call_input_spacing;
-mod command_section_indentation;
+mod mixed_indentation;
 mod comment_whitespace;
 mod concise_input;
 mod consistent_newlines;
@@ -45,7 +45,7 @@ mod version_statement_formatted;
 mod whitespace;
 
 pub use call_input_spacing::*;
-pub use command_section_indentation::*;
+pub use mixed_indentation::*;
 pub use comment_whitespace::*;
 pub use concise_input::*;
 pub use consistent_newlines::*;